add 'fee' input field on add liquidity page

+++ src/TeddySwap.UI/Models/Pool.cs
@@ -3,6 +3,6 @@ namespace TeddySwap.UI.Models;
 public record Pool
 {
     public TokenPair Pair { get; set; } = new();
-    public decimal Fee { get; set; }
+    public double Fee { get; set; }
     public decimal Tvl { get; set; }
 }
